# api/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: Fundly LOS API
  version: 1.0.0-p0
  summary: 'Loan origination system for Nigerian commercial banks: Phase 0 foundation.'
  description: 'Design-first contract for /api/v1 (TRD §10). Conventions:


    - **Auth**: staff use a same-origin Sanctum session cookie (+ XSRF token); service and partner principals use bearer tokens whose abilities
    can only narrow their role assignments.

    - **Idempotency-Key** is required on every POST that creates or causes an effect (`x-idempotency-key: required`). A replay returns the stored
    response (header `Idempotent-Replayed: true`); the same key with a different body returns 409 `idempotency-key-reused`.

    - **Concurrency**: resources carry `ETag`; PATCH requires `If-Match` (428 if missing, 412 if stale).

    - **Errors**: RFC 9457 `application/problem+json` with `type` (URN), `code`, `correlation_id`, and `errors` for validation failures.

    - **Lists**: cursor pagination with `page[size]` (≤ 100) and `page[after]`; scope filtering is always applied.

    - **Tracing**: `X-Correlation-Id` is accepted or generated and always echoed.

    - **Maker-checker** actions answer 202 with the change request; nothing changes until a different user approves it.

    - `x-permission` names the permission checked at the route; `x-step-up` marks operations needing a re-authentication within 5 minutes.

    '
  license:
    name: Proprietary (Atheris Limited)
    identifier: LicenseRef-Atheris-Proprietary
servers:
...
paths:
...
  /api/v1/parties:
    get:
      summary: 'List parties in scope'
      operationId: listParties
      tags:
        - Parties
      x-permission: 'application:view'
      parameters:
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/PageAfter'
        - name: 'filter[type]'
          in: query
          required: false
          schema:
            type: string
          description: 'individual | limited_company'
        - name: 'filter[q]'
          in: query
          required: false
          schema:
            type: string
          description: 'Name (fuzzy) or registration number'
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PartySummary'
                  meta:
                    $ref: '#/components/schemas/PageMeta'
                required:
                  - data
                  - meta
        default:
          $ref: '#/components/responses/Problem'
      security:
        - sessionCookie: []
        - bearerToken: []
    post:
      summary: 'Create a party (individual or limited company)'
      operationId: createParty
      tags:
        - Parties
      x-permission: 'party:manage'
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartyCreate'
      responses:
        201:
          description: 'Created; possible_duplicates lists intake dedupe candidates (FR-CHN-007)'
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    allOf:
                      - $ref: '#/components/schemas/Party'
                      - type: object
                        properties:
                          possible_duplicates:
                            type: array
                            items:
                              $ref: '#/components/schemas/PartyMatch'
                        required:
                          - possible_duplicates
                required:
                  - data
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
        default:
          $ref: '#/components/responses/Problem'
      security:
        - sessionCookie: []
        - bearerToken: []
      x-idempotency-key: required
...
  '/api/v1/parties/{id}/relationships':
    get:
      summary: 'Directors, shareholders, signatories and look-through beneficial owners'
      operationId: listPartyRelationships
      tags:
        - Parties
      x-permission: 'application:view'
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      relationships:
                        type: array
                        items:
                          $ref: '#/components/schemas/PartyRelationship'
                      beneficial_owners:
                        type: array
                        items:
                          $ref: '#/components/schemas/BeneficialOwner'
                    required:
                      - relationships
                      - beneficial_owners
                required:
                  - data
        default:
          $ref: '#/components/responses/Problem'
      security:
        - sessionCookie: []
        - bearerToken: []
    post:
      summary: 'Record a relationship on a company'
      operationId: addPartyRelationship
      tags:
        - Parties
      x-permission: 'party:manage'
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                related_party_id:
                  type: string
                  format: uuid
                role:
                  type: string
                  enum:
                    - director
                    - shareholder
                    - beneficial_owner
                    - signatory
                    - company_secretary
                ownership_percent:
                  type:
                    - string
                    - 'null'
                notes:
                  type:
                    - string
                    - 'null'
              required:
                - related_party_id
                - role
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/PartyRelationship'
                required:
                  - data
        default:
          $ref: '#/components/responses/Problem'
      security:
        - sessionCookie: []
        - bearerToken: []
      x-idempotency-key: required
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-fundly_session
      description: Sanctum stateful SPA session (HttpOnly, Secure, SameSite=Strict) + X-XSRF-TOKEN header on unsafe methods.
    bearerToken:
      type: http
      scheme: bearer
      description: Sanctum personal access token for service and partner accounts.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        pattern: ^[A-Za-z0-9._:-]{8,255}$
...
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      schema:
        type: string
        pattern: ^[A-Za-z0-9._:-]{8,64}$
    PageSize:
      name: page[size]
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    PageAfter:
      name: page[after]
      in: query
      schema:
        type: string
      description: Opaque cursor from meta.page.next_cursor
  headers:
    ETag:
      schema:
        type: string
      description: Entity tag for If-Match
  responses:
    Problem:
      description: Error (RFC 9457)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
        correlation_id:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
      required:
      - type
      - title
      - status
      - detail
      - code
      - correlation_id
      description: RFC 9457 problem details. `type` is a URN from the problem catalogue.
    PageMeta:
      type: object
      properties:
        page:
          type: object
          properties:
            size:
              type: integer
            next_cursor:
              type:
              - string
              - 'null'
            has_more:
              type: boolean
          required:
          - size
          - next_cursor
          - has_more
      required:
      - page
...
    PartySummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum:
            - individual
            - limited_company
        display_name:
          type: string
        org_unit_id:
          type:
            - string
            - 'null'
          format: uuid
        status:
          type: string
        registration_number:
          type:
            - string
            - 'null'
        created_at:
          type: string
          format: date-time
      required:
        - id
        - type
        - display_name
        - org_unit_id
        - status
        - registration_number
        - created_at
    Party:
      allOf:
        - $ref: '#/components/schemas/PartySummary'
        - type: object
          properties:
            cba_customer_id:
              type:
                - string
                - 'null'
            first_name:
              type:
                - string
                - 'null'
            middle_name:
              type:
                - string
                - 'null'
            last_name:
              type:
                - string
                - 'null'
            gender:
              type:
                - string
                - 'null'
            date_of_birth_masked:
              type:
                - string
                - 'null'
            nationality:
              type:
                - string
                - 'null'
            incorporation_date:
              type:
                - string
                - 'null'
            sector:
              type:
                - string
                - 'null'
            phone_masked:
              type:
                - string
                - 'null'
            email_masked:
              type:
                - string
                - 'null'
            tin_masked:
              type:
                - string
                - 'null'
            address:
              type:
                - object
                - 'null'
            identities:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  type:
                    type: string
                  value_masked:
                    type: string
                  verification_status:
                    type: string
                    enum:
                      - unverified
                      - verified
                      - failed
                  verified_at:
                    type:
                      - string
                      - 'null'
                    format: date-time
                  provider:
                    type:
                      - string
                      - 'null'
                  match_score:
                    type:
                      - string
                      - 'null'
                required:
                  - id
                  - type
                  - value_masked
                  - verification_status
            updated_at:
              type: string
              format: date-time
          required:
            - identities
            - phone_masked
            - email_masked
    PartyCreate:
      type: object
      properties:
        type:
          type: string
          enum:
            - individual
            - limited_company
        org_unit_id:
          type:
            - string
            - 'null'
          format: uuid
        first_name:
          type:
            - string
            - 'null'
        middle_name:
          type:
            - string
            - 'null'
        last_name:
          type:
            - string
            - 'null'
        gender:
          type:
            - string
            - 'null'
        date_of_birth:
          type:
            - string
            - 'null'
          format: date
        nationality:
          type:
            - string
            - 'null'
        company_name:
          type:
            - string
            - 'null'
        registration_number:
          type:
            - string
            - 'null'
        incorporation_date:
          type:
            - string
            - 'null'
          format: date
        sector:
          type:
            - string
            - 'null'
        phone:
          type:
            - string
            - 'null'
        email:
          type:
            - string
            - 'null'
        tin:
          type:
            - string
            - 'null'
        cba_customer_id:
          type:
            - string
            - 'null'
        address:
          type:
            - object
            - 'null'
          properties:
            line1:
              type: string
            city:
              type: string
            state:
              type: string
            country:
              type: string
        identities:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
                enum:
                  - bvn
                  - nin
                  - passport
                  - drivers_licence
                  - voters_card
              value:
                type: string
            required:
              - type
              - value
      required:
        - type
    PartyMatch:
      type: object
      properties:
        party_id:
          type: string
          format: uuid
        display_name:
          type: string
        type:
          type: string
        matched_fields:
          type: array
          items:
            type: string
        name_similarity:
          type:
            - string
            - 'null'
      required:
        - party_id
        - display_name
        - type
        - matched_fields
        - name_similarity
    PartyRelationship:
      type: object
      properties:
        id:
          type: string
          format: uuid
        role:
          type: string
        ownership_percent:
          type:
            - string
            - 'null'
        notes:
          type:
            - string
            - 'null'
        party:
          type: object
          properties:
            id:
              type: string
              format: uuid
            type:
              type: string
            display_name:
              type: string
          required:
            - id
            - type
            - display_name
        created_at:
          type: string
          format: date-time
      required:
        - id
        - role
        - ownership_percent
        - party
        - created_at
    BeneficialOwner:
      type: object
      properties:
        party_id:
          type: string
          format: uuid
        display_name:
          type: string
        effective_percent:
          type: string
        source:
          type: string
          enum:
            - look_through
            - declared
      required:
        - party_id
        - display_name
        - effective_percent
        - source
...